fixed some formatting, wrong but a lot closer, right now just creates new chart doesnt override the current one
