+++ index.html
@@ -38,13 +38,22 @@
           margin: 20px 0;
           font-size: 1.5em;
           color: #063806;
-        
         }
+        
+        h3{
+            text-align: center;
+            margin: 20px 0;
+            font-size: 1.5em;
+            color: #063806;
+            margin-top: 50px;
+          
+          }
 
         #description {
             text-align: center;
             margin-bottom: 20px;
         }
+
         #impact{
             text-align: center;
             margin: 40px;
@@ -79,16 +88,44 @@
             max-width: 48%;
     
         }
-        #map-container {
-            margin-top: 100px;
-        }
 
-        #map {
+
+        #map-container {
+            display: flex;
+            margin-top: 100px;
+            justify-content: flex-start; /* Align items horizontally */
+            align-items: flex-start;    /* Align items vertically */
+            justify-content: space-between;
+            align-items: center; /* Aligns items (map and legend) vertically */
+            gap: 0px;
             margin: 20px auto;
+            max-width: 1200px;
+            gap: 20px;
+            position: relative; /* Allows positioning of the legend relative to the container */
+        }
+        
+        #map {
+            flex: 5; /* Take up more space compared to the legend */
+            justify-content: right; /* Centers content horizontally */
+            align-items: center; /* Centers content vertically */
             width: 800px;
             height: 500px;
-            border: 1px solid #ccc;
+            margin: 20px auto;
+            margin-left: 150px; /* Pushes the entire container to the right */
         }
+        
+        #legend-container {
+            flex: 1; /* Take up less space */
+            min-width: 10px;
+            padding: 0px;
+            background-color: #FFF;
+            align-items: center; 
+            position: absolute; /* Positions the legend relative to #map-container */
+            right: -50px; 
+            top: 50%; /* Aligns the legend vertically */
+            transform: translateY(-50%); /* Centers the legend vertically */
+        }
+    
 
         #controls {
             display: flex;
@@ -145,15 +182,16 @@
         </div>
     </div>
 
+    <h3>Percent Change (kg/cap) By Country</h3>
     <div id="map-container">
-        <div id="map"></div>
-        <div id="controls">
-            <div class="controls-right"></div>
-        </div>
+        
+            <div id="map"></div>
+
+            <div id="legend-container" /div>
     </div>
 
     <script src="pie.js"></script>
-    <script src="chart.js"></script>
+    <script type = "module" src="chart.js"></script>
     <script type="module" src="map.js"></script>
 </body>
 </html>
